Backend/requirements.txt
tensorflow-cpu
deepface
numpy
pillow
